separate out Render.Components.TimingView module
